refactor: use standardized folder structure
