--- libs/journeys/ui/src/components/Video/VideoControls/Desktop/index.ts
@@ -1,16 +1,0 @@
-import { FullscreenButton } from '../components/FullscreenButton'
-import { TimeLabel } from '../components/TimeLabel'
-
-import { PlayButton } from './PlayButton'
-import { Progress } from './Progress'
-import { Root } from './Root'
-import { VolumeControls } from './VolumeControls'
-
-export const Desktop = {
-  Root,
-  PlayButton,
-  TimeLabel,
-  Progress,
-  VolumeControls,
-  FullscreenButton
-}
