+++ libs/journeys/ui/src/components/Video/VideoControls/DesktopControls/index.ts
@@ -1,0 +1,1 @@
+export { DesktopControls } from './DesktopControls'
