--- libs/journeys/ui/src/components/Video/VideoControls/Mobile/index.ts
@@ -1,12 +1,0 @@
-import { FullscreenButton } from '../components/FullscreenButton'
-import { TimeLabel } from '../components/TimeLabel'
-
-import { Progress } from './Progress'
-import { Root } from './Root'
-
-export const Mobile = {
-  Root,
-  TimeLabel,
-  Progress,
-  FullscreenButton
-}
