+++ libs/journeys/ui/src/components/Video/VideoControls/MobileControls/index.ts
@@ -1,0 +1,1 @@
+export { MobileControls } from './MobileControls'
