+++ libs/journeys/ui/src/components/Video/VideoControls/PlaybackIcon/index.ts
@@ -1,0 +1,1 @@
+export { PlaybackIcon } from './PlaybackIcon'
