--- libs/journeys/ui/src/components/Video/VideoControls/components/TimeLabel.tsx
@@ -1,21 +1,0 @@
-import Typography from '@mui/material/Typography'
-
-export function TimeLabel({
-  displayTime,
-  duration
-}: {
-  displayTime: string
-  duration: string
-}): JSX.Element {
-  return (
-    <Typography
-      variant="caption"
-      color="secondary.main"
-      noWrap
-      overflow="unset"
-      sx={{ p: 2 }}
-    >
-      {displayTime} / {duration}
-    </Typography>
-  )
-}
