+++ libs/journeys/ui/src/components/Video/VideoControls/utils/PlaybackReducer/index.ts
@@ -1,0 +1,1 @@
+export { playbackReducer, PlaybackState } from './playbackReducer'
